refactor: put the CLI creation functions into a dedicated file to improve future testing
